Refactor port scanner implementation: move PortScanner to tcp.py and update imports
