+++ tests/unit/test_tcp_scanner.py
@@ -1,9 +1,9 @@
 from unittest.mock import MagicMock, patch
 
-from scanner.core.scanner import PortScanner
+from scanner.core.tcp import PortScanner
 
 
-@patch("scanner.core.scanner.socket.socket")
+@patch("scanner.core.tcp.socket.socket")
 def test_scan_ports_open(mock_socket_class):
     mock_socket = MagicMock()
     mock_socket.connect_ex.return_value = 0
@@ -17,7 +17,7 @@ def test_scan_ports_open(mock_socket_class):
     assert result["scan_results"][0]["port"] == 80
 
 
-@patch("scanner.core.scanner.socket.socket")
+@patch("scanner.core.tcp.socket.socket") 
 def test_scan_ports_closed(mock_socket_class):
     mock_socket = MagicMock()
     mock_socket.connect_ex.return_value = 1
